backend/requirements.txt
fastapi
uvicorn
torch
transformers
gtts
pillow
